Cleaned up testing and changed links
The testing frameworkneeded some love, including automated setup, clean
locations, refactoring the code a bit to not repeat the same thing over
and over, etc. The tests also needed to be updated to use Thali as a
NPM style require not a file require.

Also updated the links to all point directly at story 0.
