+++ test/www/index.html
@@ -7,7 +7,6 @@
     <title>Thali Scenario Tests</title>
 
     <script src="cordova.js"></script>
-    <script src="jxcore.js"></script>
     <script src="js/thali_main.js"></script>
 </head>
 <body>
