+++ test/www/jxcore/mockmobile.js
