--- test/www/jxcore/test/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "thali-cordova-plugin-test",
-  "version": "1.0.0",
-  "description": "Thali Cordova Plugin Test",
-  "main": "thaliscenarios.js",
-  "scripts": {
-    "test": "tape thaliscenarios.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/thaliproject/Thali_CordovaPlugin.git"
-  },
-  "keywords": [
-    "Thali",
-    "PouchDB",
-    "Android",
-    "iOS"
-  ],
-  "author": "Microsoft Corporation",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/thaliproject/Thali_CordovaPlugin/issues"
-  },
-  "homepage": "https://github.com/thaliproject/Thali_CordovaPlugin#readme",
-  "devDependencies": {
-    "randomstring": "^1.0.7",
-    "tape": "^4.0.0"
-  },
-  "dependencies": {
-    "express": "4.13.0"
-  }
-}
