+++ thali/install/package.json
@@ -9,7 +9,7 @@
     "fs-extra-promise":"^0.2.0"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "./setUpTests.sh"
   },
   "author": "Microsoft",
   "license": "MIT"
